Remove pull request job; update cron day
The fix for the template works locally. We should see this
successfully run sometime tomorrow. The CI flow got to the
point of reproducing the failure. It reproduces the failure
because it's pulling the template from the master branch.
The intent is to merge this branch into master, then see the
job pass.